backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
drf-yasg==1.21.7
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-storages==1.14.2
django-rest-auth==0.9.5